% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{debug}
\alias{debug}
\title{Debug information}
\usage{
debug()
}
\value{
Returns nothing.
}
\description{
Debug information
}
\examples{
debug()

## This gives an error if _R_CHECK_DEPENDS_ONLY_EXAMPLES_=true
library(listenv)

}
