% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreach_it.R
\name{foreach_it}
\alias{foreach_it}
\title{Iterate over a Vector using Foreach}
\usage{
foreach_it(X)
}
\arguments{
\item{X}{A vector of length N}
}
\value{
A vector of length N
}
\description{
Iterate over a Vector using Foreach
}
