% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Displays A Message}
\usage{
hello(test = FALSE)
}
\arguments{
\item{test}{If TRUE, globals are queried.}
}
\value{
Returns nothing.
}
\description{
A function that says hello to the world.
}
\examples{
hello()

}
