% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_variance.R
\name{mean_and_variance}
\alias{mean_and_variance}
\title{Calculate the Mean and Variance}
\usage{
mean_and_variance(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A named numeric vector of length two
with elements `mean` and `variance`.
}
\description{
Calculate the Mean and Variance
}
