% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Person.R
\name{Person}
\alias{Person}
\title{A Person}
\usage{
Person(name = "Alice", ...)
}
\arguments{
\item{name}{The name of the person.}

\item{\ldots}{(optional) Additional named elements.}
}
\value{
Returns an [R.oo::Object] of class Person.
}
\description{
Create a Person
}
\examples{
alice <- Person("Alice")
bob <- Person("Bob")

hello(alice)
hello(bob)
}
