% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello.default}
\alias{hello.default}
\alias{hello}
\title{Say Hello to the World}
\usage{
\method{hello}{default}(x = NULL, ...)
}
\arguments{
\item{x}{An \R object}

\item{\ldots}{Not used.}
}
\value{
Returns (invisibly) the object `x`.
}
\description{
A function that says hello to the world.
}
\examples{
hello()
hello("Alice")
hello("Bob")
}
