% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloppy_if.R
\name{sloppy_if}
\alias{sloppy_if}
\title{Test a Condition}
\usage{
sloppy_if(cond)
}
\arguments{
\item{cond}{(logical)}
}
\value{
Returns TRUE if `cond` is TRUE, otherwise FALSE.
}
\description{
Test a Condition
}
\examples{
library(teeny)

value <- Sys.getenv("_R_CHECK_LENGTH_1_CONDITION_")
message("_R_CHECK_LENGTH_1_CONDITION_=", value)
do_validate <- isTRUE(as.logical(value))

## This will give an error if _R_CHECK_LENGTH_1_CONDITION_=true
res <- tryCatch({
  if(c(TRUE, TRUE)) message("TRUE") else message("FALSE")
}, error = function(e) e)
print(res)
stopifnot(getRversion() < "3.4.0" || inherits(res, "error") ||
  is.element(value, c("", "false", "warn")))


## This will give an error if _R_CHECK_LENGTH_1_CONDITION_=true
res <- tryCatch({
  if(c(FALSE, TRUE)) message("TRUE") else message("FALSE")
}, error = function(e) e)
print(res)
stopifnot(getRversion() < "3.4.0" || inherits(res, "error") ||
  is.element(value, c("", "false", "warn")))

## This will give an error if _R_CHECK_LENGTH_1_CONDITION_=true
res <- tryCatch({
  sloppy_if(c(TRUE, TRUE))
}, error = function(e) e)
print(res)
stopifnot(getRversion() < "3.4.0" || inherits(res, "error") ||
  is.element(value, c("", "false", "warn")))

## This will give an error if _R_CHECK_LENGTH_1_CONDITION_=true
res <- tryCatch({
  sloppy_if(c(TRUE, FALSE))
}, error = function(e) e)
print(res)
stopifnot(getRversion() < "3.4.0" || inherits(res, "error") ||
  is.element(value, c("", "false", "warn")))
}
